refactor: AbstractFareCalculatorChain 로 로직 공통화
