+++ src/main/java/nextstep/subway/domain/fare/DefaultFareCalculator.java
@@ -2,22 +2,9 @@ package nextstep.subway.domain.fare;
 
 import nextstep.subway.domain.Path;
 
-import java.util.Objects;
-
-public class DefaultFareCalculator implements FareCalculatorChain {
+public class DefaultFareCalculator extends AbstractFareCalculatorChain {
 
     private static final int MINIMUM_FARE = 1250;
-
-    private FareCalculatorChain nextChain;
-
-    @Override
-    public int calculate(Path path, int initialFare) {
-        if (Objects.isNull(nextChain)) {
-            return MINIMUM_FARE;
-        }
-
-        return nextChain.calculate(path, MINIMUM_FARE);
-    }
 
     @Override
     public boolean support(Path path) {
@@ -25,8 +12,8 @@ public class DefaultFareCalculator implements FareCalculatorChain {
     }
 
     @Override
-    public void setNextChain(FareCalculatorChain chain) {
-        this.nextChain = chain;
+    protected int convert(Path path, int initialFare) {
+        return MINIMUM_FARE;
     }
 
 }
