+++ src/main/java/nextstep/subway/domain/fare/FirstRangeFareCalculator.java
@@ -2,32 +2,12 @@ package nextstep.subway.domain.fare;
 
 import nextstep.subway.domain.Path;
 
-import java.util.Objects;
-
-public class FirstRangeFareCalculator implements FareCalculatorChain {
+public class FirstRangeFareCalculator extends AbstractFareCalculatorChain {
 
     private static final int FARE_PER_UNIT_DISTANCE = 100;
     private static final float FIRST_UNIT_DISTANCE = 5f;
     private static final int RANGE_START = 10;
     private static final int RANGE_ENDS = 50;
-
-    private FareCalculatorChain nextChain;
-
-    @Override
-    public int calculate(Path path, int initialFare) {
-        if (!support(path)) {
-            return initialFare;
-        }
-
-        var targetDistance = getTargetDistance(path.extractDistance());
-        var fare =  initialFare + getFirstRangeOverFare(targetDistance);
-
-        if (Objects.isNull(nextChain)) {
-            return fare;
-        }
-
-        return nextChain.calculate(path, fare);
-    }
 
     @Override
     public boolean support(Path path) {
@@ -35,8 +15,9 @@ public class FirstRangeFareCalculator implements FareCalculatorChain {
     }
 
     @Override
-    public void setNextChain(FareCalculatorChain chain) {
-        this.nextChain = chain;
+    protected int convert(Path path, int initialFare) {
+        var targetDistance = getTargetDistance(path.extractDistance());
+        return initialFare + getFirstRangeOverFare(targetDistance);
     }
 
     private int getTargetDistance(int distance) {
