+++ src/main/java/nextstep/subway/domain/fare/SecondRangeFareCalculator.java
@@ -2,30 +2,11 @@ package nextstep.subway.domain.fare;
 
 import nextstep.subway.domain.Path;
 
-import java.util.Objects;
-
-public class SecondRangeFareCalculator implements FareCalculatorChain {
+public class SecondRangeFareCalculator extends AbstractFareCalculatorChain {
 
     private static final int FARE_PER_UNIT_DISTANCE = 100;
     private static final int RANGE_START = 50;
     private static final float SECOND_UNIT_DISTANCE = 8f;
-
-    private FareCalculatorChain nextChain;
-
-    @Override
-    public int calculate(Path path, int initialFare) {
-        if (!support(path)) {
-            return initialFare;
-        }
-
-        var fare = initialFare + getSecondRangeOverFare(getTargetDistance(path.extractDistance()));
-
-        if (Objects.isNull(nextChain)) {
-            return fare;
-        }
-
-        return nextChain.calculate(path, fare);
-    }
 
     @Override
     public boolean support(Path path) {
@@ -33,8 +14,8 @@ public class SecondRangeFareCalculator implements FareCalculatorChain {
     }
 
     @Override
-    public void setNextChain(FareCalculatorChain chain) {
-        this.nextChain = chain;
+    protected int convert(Path path, int initialFare) {
+        return initialFare + getSecondRangeOverFare(getTargetDistance(path.extractDistance()));
     }
 
     private int getTargetDistance(int distance) {
